GTP, GTPv2: stat taps can't be wmem_file_scope
Stat tap windows can be opened by the GUI (e.g., a
ServiceResponseTimeDialog) when no file is open, and persist
past a file being closed, i.e. outside of wmem_file_scope().

Items concerning the taps should not be created in wmem_file_scope().
This fails an assert, which crashes when built for a Debug target.

To use wmem, we would need to create a scope appropriate for the lifetime
of the ServiceReponseTimeDialog or other Tap dialog (or else add a
callback mechanism to srt table to free items created in epan scope.)

Partially revert 47b310da470c6dc527a63f2ee2b4bbafafd5e290
(the part where the stat taps are concerned.)

Related to #19620