WOW: Fix use of unitialized memory
tvb_get_raw_bytes_as_string doesn't check lengths, because it's
used elsewhere when the length is unknown. If we use
tvb_get_string_enc, that checks the offsets and throws an
exception as appropriate, but then we have to use g_utf8_strreverse
due to the possibility of UTF-8 REPLACEMENT CHARACTERs.

To handle embedded nulls properly, we need to be using counted
strings (like wmem_strbuf_t) in more places.

Fix #19621