epan: When exporting as C arrays, don't end a line with backslash
If the last hex character in a line of a packet is 0x5c (ASCII '\'),
we have to make sure that doesn't end the line, because a backslash-newline
gets treated as a spliced line in C (and that happens in syntax
translation in a very early phase, before comment removal).

We can't just add whitespace (even though §5.1.2.2 of the C standard
says we could), because gcc (and clang) helpfully assume that a
backslash with only whitespace before a newline is probably programmer
error and treats it as a continuation while warning about it:
https://gcc.gnu.org/onlinedocs/gcc/Escaped-Newlines.html#Escaped-Newlines

Surround the text with `|` because that's what hexdump(1) does, though
really most anything that isn't whitespace would do.

Fix #19615