Qt: Only save the filtered rows to the Map
Be a little more careful in TrafficTab about whether we
want the Proxy Model or the Source Model, since the tap
filter doesn't exclude rows from the source model (since
the non-displayed rows are used to as part of the percentages
of the total.)

When we want to respect the display filtering or the sorting,
use the proxy model.

In the tap data model, only claim we have GeoIPData (which makes
the map available) if a row that isn't filtered out has data.

Fix #18530