X11: static-ify the generated dissector.
Many of the generated "const" fields can be made static.  For local variables,
this means that the compiler will no longer emit code to allocate the structure
on the stack and initialize it upon each function entry -- 29 kBytes of code in
total gets removed by this change -- which is surely a performance win when
dissecting this protocol.

Interestingly, libwireshark.so ends up growing overall due to the addition of
52 kBytes of relocation entries to patch up references within the relocatable
read-only data section.  Clearly the generated code was more space-efficient
than the relocation entries...