Qt: Add width and alignment to column preferences
Add the ability to change the width and alignment of columns
from the Column Preferences.

This also makes it easier to eventually fix #15529 by having all
the column-relevant details edited at once. In order to properly
solve that issue, the column indices from the preferences and the
recent settings need to be kept in sync, instead of using the format
as the unique key.

Related to #15529