recent: Don't use format info at all
We don't need to read or use the column format string in
the recent settings column width info, because now they're
in sync with the indices used in prefs.col_list.

Continue to write it when writing out the recent settings for
backwards compatibility with older version.

We can also remove the workaround for #14177 from
commit c62dadd31f859a60b970f699992a5ddba837a42c as it's
not needed.

This solves the width and alignment part of #15529. Hidden
states are still to do.