socketcan: clean up some variable names.
The "flags" field in CAN classic and CAN FD is relly a "flags and ID"
field; rename appropriatey.

Rename can_..._flags_id_fd fields to canfd_..._flags_id, so as not to
mix up "flags_id" and "fd" - the latter refers to CAN FD, the former
refers to the two values in the item.