TCP: Fix location of Timestamp Option values
When syn cookie support was added, a new tree was added under
the TCP timestamp option TSVal. Unfortunately, it was done
so by reusing the proto_item pointer that was pointing to the
tree item for the top level of the option, before it was done
having text appended to it.

Add a new proto_item pointer so that the text gets appended
to the original place.

Fixup acc0260c84fccb4049ac569e72aaf6a6899ef77a