blf: if there are no log containers, return FALSE when pulling one.
blf_pull_next_logcontainer() shuld check, aftre calling
blf_find_next_logcontainer() to see if the size of the
GArray of log containers is 0; if so, that means no log container was
found, probably because there aren't any to find.

This fixes a case where attempting to read a file with no log containers
causes a crash.