socketcan: don't use the byte-swapping preference for CAN XL.
For WTAP_ENCAP_SOCKETCAN(LINKTYPE_CAN_SOCKETCAN), the only time any CAN
XL header would need byte-swapping would be with captures fom a
big-endian machine with current (but not upcoming) versions of libpcap.
Most captures are *probably* going to be on little-endian machines, so
that's probably not very likey - and the current byte-swapping
preference means "the ID/flags field is in *little-endian* byte order
and was done with a libpcap that didn't put it in big-endian byte
order", which means that, if it's necessary, it would be a separate
preferece.  We'll add such a proference if it's ever necessary.

For dissect_socketcan_classic() and dissect_socketcan_fd(), those
dissectors probably shouldn't need to exist; the libwiretap modules that
use them should be changed to use WTAP_ENCAP_SOCKETCAN.  In any case,
they're not used for CAN XL, so the CAN XL encapsulation argument they
pass is irrelevant.

For dissect_socketcan_sll(), the SocketCAN header is in *host* byte
order, with libpcap and libwireshark doing the usual byte-swapping dance
when reading a pcap or pcapng packet from a file or file section in the
opposite byte order from the machine reading the file (that needs to be
updated to handle CAN XL, and that's both a libpcap issue and a
libwiretap issue, each of which I'll address).

So, for now, don't use the byte-swapping preference to control the
encapsulatio for CAN XL.

Rename the prefrence to indicate what it does, namely control whether
the ID/flags field in the CAN classic/CAN FD header willl be
byte-swapped (and, by implication, what it *doesn't* do, namely control
whether fields in the CAN XL are byte-swapped).