dfilter: Don't allow "matches" operator with FT_FRAMENUM
The matches operator implicitly converts non-stringlike fields
that have value strings to their value string value. (This is
not the same as the string representation of the number, which
applying the string function first would do, but it usually less
useful and worse performance than using numeric comparisons.)

However, FT_FRAMENUM fields have a hfinfo->strings but it is not
strings used for conversion, it is an overload with the special
ft_framenum_type_t, so don't convert.

This prevents a segmentation fault if expressions with
expressions like 'gtp.response_in ~ "test"'