Qt: Add case sensitive find to Follow Stream and Show Packet Bytes
Add a check box for case sensitivity when finding in follow stream
and show packet bytes.

Note that QPlainTextEdit::find() has a bit unexpected behavior with
QRegularExpression (https://bugreports.qt.io/browse/QTBUG-88721).
Searches are case-insensitive by default there too, respecting
the default options. This is a change from the older QRegExp, where
the option to find was ignored, and only the regex option was used,
so for the last few releases regexp searches have been case-insensitive
as well by default. (?-i) has been available to mode switch.

We might want to move the various keyboard handling from
FollowStreamDialog and ShowPacketBytesDialog and instead have
FindLineEdit install shortcuts on the parents when constructed.
That would be a little cleaner separation.

We might also want to move the buttons and label to a separate
composite widget class, that signals the parent to start a
find.

Fix #3784