prefs: Don't remove initial handles from auto port prefs
When removing old values from a dissector table via changing an
an auto port preference associated with Decode As, set the current
handle to NULL instead of deleting the entry. This avoids removing
the default value for ports that have them, and makes it possible
for users to see the original handle in Decode As.

When setting a dissector table entry to NULL, if it doesn't have
an initial handle value, remove the entry from the table.