gRPC: Add separate dissector handles for gRPC Web and Web Text
Rather than registering one gRPC dissector handle to three
flavors of gRPC and then figuring out which one got called
by inspecting pinfo->match_string, register different dissectors
with descriptions to each type, so we know which dissector was
called.

This especially avoids assertions when a dissector is set via
Decode As, or via the stream ID table (even though the stream
ID table won't put the gRPC dissector into the special
streaming "reassembly as you go" mode, instead falling back
to the standard HTTP/2 reassemble at the end mode.)

Also, if we have the (bidirectional) streaming mode set for
the stream ID but we don't have a streaming subdissector handle
for the (unidirectional) content type, set it to the data handle
to avoid an assertion. We could possibly try to set it to the
content type used in the other direction, but that has potential
issues.

Fix #19679