NFS: Add a NFSv4 SRT table for operations
For NFSv4, the ONC-RPC procedure is always either NULL or COMPOUND.
COMPOUND procedures have multiple operations, which is more interesting
for analysis.

Add a NFSv4 tap and a SRT table for the individual NFSv4 opcodes.
Have a table that tracks each opcode, and a table for those considered
the main opcode in a compound procedure. Also have a table for NFSv4
callback.

Fix #1693