Qt: Don't allow tap dialogs to apply with no capture file
tap dialogs disable the edit and apply button after the
capture file is closed. Also check if the capture file is
valid ("open, readable, and tappable") and disable the buttons
in that case as well. We don't get the closed signal if the
dialog is initially opened with no capture file. Perform the
check upon dialog creation.

This prevents some heap-use-after-free and garbage entries in
the tap (e.g., SRT) tables if the Apply button is pushed when
the dialog was opened with no capture file opened.

We could (should?) disable the various tap dialog menu items
elsewhere too (WiresharkMainWindow::setMenusForCaptureFile ?)
when no capture file is open.