opus: Fix dissection of payloads with multiple Opus frames
In VBR packets with a variable number of Opus frames, only the
first N - 1 (N must be > 0) frames have their frame size explicitly
included. The last frame size is calculated from the remaining bytes
(also removing the padding.)

Fix a few other locations of checks.

Add the frame size, which is a special kind of varint, as an integer
instead of bytes.

Fix #19664