Goto the nearest undisplayed frame
Add a parameter to cf_goto_frame so that if it is called
with an undisplayed frame, we can optionally go to the nearest
displayed frame, if there is one.

For following FT_FRAMENUM links, continue to fail if the
requested frame isn't displayed, because we probably want
the exact frame in that case. (There is "Go to Previous
Packet in History" so it wouldn't be terrible to go there
anyway.)

We search for the next displayed packet and go there if it's
closer in number than the previous displayed packet.
That is slightly overkill; we could just go to the previously
displayed frame regardless, but the O(log n) search to see if
the the next displayed frame is closer shouldn't be too bad.

The frame data prev_dis_num currently means "previous frame
with a timestamp" because that's what it's used for (except
in the wireless timeline, which does something similar to this
already.) So for now this will ignore frames without a timestamp,
though that will merely fall back to the present behavior.

Fix #2988